requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.6.1
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-multipart==0.0.9
